Create Example 1
In a school there are some teachers who handle two subjects (Maths and English). When
the feedback was taken their feedback was present in both Maths Feedback as well as English Feedback.
Write a method to create consolidated feedback for the teachers for English and Maths. For those
taking both the subjects, the highest feedback is taken. Write a method to accept two Dictionary and return a
Dictionary object containing the feedbacks of all teachers in maths and English